pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-crawler"
version = "0.1.0"
description = "Web crawler for books.toscrape.com"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "selectolax>=0.3.20",
    "pydantic>=2.5.0",
    "motor>=3.3.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "apscheduler>=3.10.0",
    "slowapi>=0.1.6"
]

[tool.pytest.ini_options]
pythonpath = ["."]